pyproject.toml
[build-system]
requires = ["poetry-core>=1.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "warifuri"
version = "0.2.0"
description = "Task automation with cross-project dependency resolution"
authors = ["f43a9a"]
readme = "README.md"
packages = [{include = "warifuri", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
PyYAML = ">=6.0"
gitpython = ">=3.1.0"
requests = ">=2.28.0"
flask = ">=2.0.0"
click = ">=8.0.0"

[tool.poetry.scripts]
warifuri = "warifuri.__main__:main"

[tool.poetry.group.dev.dependencies]
mypy = "*"
ruff = "*"
pytest = "*"

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = false
strict_equality = true
exclude = [
  "tests/.*",
  "scripts/.*",
  ".*_test\\.py$",
]

[[tool.mypy.overrides]]
module = ["git.*", "flask.*", "requests.*", "pytest.*"]
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 100
exclude = [".git", "__pycache__", "build", "dist"]

[tool.ruff.lint]
select = ["F", "B", "E", "W"]
ignore = ["E203", "E501", "C901"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "--color=yes"
